+++ spec/sippy_cup/scenario_spec.rb
@@ -14,11 +14,6 @@ describe SippyCup::Scenario do
 
   subject(:scenario) { described_class.new 'Test', default_args.merge(args) }
 
-  it "creates a media stream on initialization" do
-    SippyCup::Media.should_receive(:new).once
-    subject
-  end
-
   it "takes a block to generate a scenario" do
     s = described_class.new 'Test', default_args do
       invite
@@ -102,7 +97,7 @@ describe SippyCup::Scenario do
     it "sends a REGISTER message" do
       subject.register 'frank'
 
-      subject.to_xml.should match(%r{<send .*>})
+      subject.to_xml.should match(%r{<send.*>})
       subject.to_xml.should match(%r{REGISTER})
     end
 
@@ -143,6 +138,7 @@ describe SippyCup::Scenario do
 
     context "when a password is provided" do
       it "expects a 401 response" do
+        pending "Need to check for initial request, then 401, then retry with authentication"
         subject.register 'frank', 'abc123'
         subject.to_xml.should match(%r{<recv response="401" auth="true" optional="false"/>})
       end
@@ -218,19 +214,19 @@ describe SippyCup::Scenario do
     it "expects a 200 with rrs and rtd true" do
       subject.receive_answer
 
-      scenario.to_xml.should match(%q{<recv response="200" rrs="true" rtd="true"/>})
+      scenario.to_xml.should match(%q{<recv response="200" rrs="true" rtd="true">})
     end
 
     it "allows passing options to the recv expectation" do
       subject.receive_answer foo: 'bar'
 
-      scenario.to_xml.should match(%q{<recv response="200" rrs="true" rtd="true" foo="bar"/>})
+      scenario.to_xml.should match(%q{<recv response="200" rrs="true" rtd="true" foo="bar">})
     end
 
     it "allows overriding options" do
       subject.receive_answer rtd: false
 
-      scenario.to_xml.should match(%q{<recv response="200" rrs="true" rtd="false"/>})
+      scenario.to_xml.should match(%q{<recv response="200" rrs="true" rtd="false">})
     end
   end
 
@@ -268,10 +264,14 @@ describe SippyCup::Scenario do
     end
 
     context "when media is present" do
-      before { subject.sleep 1 }
+      before do
+        subject.answer
+        subject.sleep 1
+      end
 
       it "starts the PCAP media" do
         subject.ack_answer
+        subject.sleep 1
         subject.to_xml(:pcap_path => "/dev/null").should match(%r{<nop>\n.*<action>\n.*<exec play_pcap_audio="/dev/null"/>\n.*</action>\n.*</nop>})
       end
     end
@@ -280,24 +280,6 @@ describe SippyCup::Scenario do
       it "does not start the PCAP media" do
         subject.ack_answer
         subject.to_xml(:pcap_path => "/dev/null").should_not match(%r{<nop>\n.*<action>\n.*<exec play_pcap_audio="/dev/null"/>\n.*</action>\n.*</nop>})
-      end
-    end
-
-    context "when a from user is specified" do
-      let(:args) { {from_user: 'frank'} }
-
-      it "includes the specified user in the From and Contact headers" do
-        subject.ack_answer
-        subject.to_xml.should match(%r{From: "frank" <sip:frank@})
-        subject.to_xml.should match(%r{Contact: <sip:frank@})
-      end
-    end
-
-    context "when no from user is specified" do
-      it "uses a default of 'sipp' in the From and Contact headers" do
-        subject.ack_answer
-        subject.to_xml.should match(%r{From: "sipp" <sip:sipp@})
-        subject.to_xml.should match(%r{Contact: <sip:sipp@})
       end
     end
   end
@@ -312,6 +294,8 @@ describe SippyCup::Scenario do
       xml.should =~ /recv response="183".*optional="true"/
       xml.should =~ /recv response="200"/
       xml.should_not =~ /recv response="200".*optional="true"/
+      xml.should match(%r{<send>})
+      xml.should match(%r{ACK})
     end
 
     it "passes through additional options" do
@@ -322,6 +306,8 @@ describe SippyCup::Scenario do
       xml.should =~ /recv .*foo="bar".*response="180"/
       xml.should =~ /recv .*foo="bar".*response="183"/
       xml.should =~ /recv .*response="200" .*foo="bar"/
+      xml.should match(%r{<send.*foo="bar".*>})
+      xml.should match(%r{ACK})
     end
   end
 
@@ -361,24 +347,6 @@ describe SippyCup::Scenario do
       subject.send_bye foo: 'bar'
       subject.to_xml.should match(%r{<send foo="bar".*>})
     end
-
-    context "when a from user is specified" do
-      let(:args) { {from_user: 'frank'} }
-
-      it "includes the specified user in the From and Contact headers" do
-        subject.send_bye
-        subject.to_xml.should match(%r{From: "frank" <sip:frank@})
-        subject.to_xml.should match(%r{Contact: <sip:frank@})
-      end
-    end
-
-    context "when no from user is specified" do
-      it "uses a default of 'sipp' in the From and Contact headers" do
-        subject.send_bye
-        subject.to_xml.should match(%r{From: "sipp" <sip:sipp@})
-        subject.to_xml.should match(%r{Contact: <sip:sipp@})
-      end
-    end
   end
 
   describe '#receive_bye' do
@@ -407,22 +375,6 @@ describe SippyCup::Scenario do
       subject.okay foo: 'bar'
       subject.to_xml.should match(%r{<send foo="bar".*>})
     end
-
-    context "when a from user is specified" do
-      let(:args) { {from_user: 'frank'} }
-
-      it "includes the specified user in the Contact header" do
-        subject.okay
-        subject.to_xml.should match(%r{Contact: <sip:frank@})
-      end
-    end
-
-    context "when no from user is specified" do
-      it "uses a default of 'sipp' in the Contact header" do
-        subject.okay
-        subject.to_xml.should match(%r{Contact: <sip:sipp@})
-      end
-    end
   end
 
   describe '#wait_for_hangup' do
@@ -434,10 +386,25 @@ describe SippyCup::Scenario do
     end
   end
 
+  describe '#call_length_repartition' do
+    it 'create a partition table' do
+      subject.call_length_repartition('1', '10', '2')
+      scenario.to_xml.should match('<CallLengthRepartition value="1,3,5,7,9"/>')
+    end
+  end
+
+  describe '#response_time_repartition' do
+    it 'create a partition table' do
+      subject.response_time_repartition('1', '10', '2')
+      scenario.to_xml.should match('<ResponseTimeRepartition value="1,3,5,7,9"/>')
+    end
+  end
+
   describe 'media-dependent operations' do
     let(:media) { double :media }
     before do
       SippyCup::Media.should_receive(:new).once.and_return media
+      scenario.ack_answer
       media.stub :<<
     end
 
@@ -485,6 +452,7 @@ describe SippyCup::Scenario do
 
   describe "#send_digits with a SIP INFO DTMF mode" do
     let(:args) { {dtmf_mode: 'info'} }
+    before { scenario.answer }
 
     it "creates the requested DTMF string as SIP INFO messages" do
       scenario.send_digits '136'
@@ -516,6 +484,7 @@ describe SippyCup::Scenario do
       end
 
       it "writes the PCAP media to disk at name.pcap" do
+        scenario.ack_answer
         scenario.send_digits '123'
 
         scenario.compile!
@@ -540,6 +509,7 @@ describe SippyCup::Scenario do
       end
 
       it "writes the PCAP media to disk at filename.pcap" do
+        scenario.ack_answer
         scenario.send_digits '123'
 
         scenario.compile!
@@ -575,7 +545,10 @@ describe SippyCup::Scenario do
     end
 
     context "with media" do
-      before { scenario.sleep 1 }
+      before do
+        scenario.ack_answer
+        scenario.sleep 1
+      end
 
       it "writes the PCAP media to a Tempfile and returns it" do
         files = scenario.to_tmpfiles
@@ -589,7 +562,6 @@ describe SippyCup::Scenario do
       end
 
       it "puts the PCAP file path into the scenario XML" do
-        scenario.ack_answer
         files = scenario.to_tmpfiles
         files[:scenario].read.should match(%r{play_pcap_audio="#{files[:media].path}"})
       end
@@ -604,7 +576,7 @@ describe SippyCup::Scenario do
 <![CDATA[
 INVITE sip:[service]@[remote_ip]:[remote_port] SIP/2.0
 Via: SIP/2.0/[transport] [local_ip]:[local_port];branch=[branch]
-From: "sipp" <sip:sipp@[local_ip]>;tag=[call_number]
+From: "sipp" <sip:sipp@[local_ip]:[local_port]>;tag=[call_number]
 To: <sip:[service]@[remote_ip]:[remote_port]>
 Call-ID: [call_id]
 CSeq: [cseq] INVITE
@@ -624,20 +596,24 @@ a=rtpmap:0 PCMU/8000
 a=rtpmap:101 telephone-event/8000
 a=fmtp:101 0-15
 ]]>
-</send>
+<action><assignstr assign_to="remote_addr" value="[service]@[remote_ip]:[remote_port]"/><assignstr assign_to="local_addr" value="sipp@[local_ip]:[local_port]"/><assignstr assign_to="call_addr" value="[service]@[remote_ip]:[remote_port]"/></action></send>
   <recv response="100" optional="true"/>
   <recv response="180" optional="true"/>
   <recv response="183" optional="true"/>
-  <recv response="200" rrs="true" rtd="true"/>
+  <recv response="200" rrs="true" rtd="true">
+    <action>
+      <ereg regexp="&lt;sip:(.*)&gt;.*;tag=([^;]*)" search_in="hdr" header="To:" assign_to="dummy,remote_addr,remote_tag"/>
+    </action>
+  </recv>
   <send>
 <![CDATA[
 ACK [next_url] SIP/2.0
 Via: SIP/2.0/[transport] [local_ip]:[local_port];branch=[branch]
-From: "sipp" <sip:sipp@[local_ip]>;tag=[call_number]
+From: "sipp" <sip:sipp@[local_ip]:[local_port]>;tag=[call_number]
 To: <sip:[service]@[remote_ip]:[remote_port]>[peer_tag_param]
 Call-ID: [call_id]
 CSeq: [cseq] ACK
-Contact: <sip:sipp@[local_ip]:[local_port];transport=[transport]>
+Contact: <sip:[$local_addr];transport=[transport]>
 Max-Forwards: 100
 User-Agent: SIPp/sippy_cup
 Content-Length: 0
@@ -653,23 +629,41 @@ SIP/2.0 200 OK
 [last_To:]
 [last_Call-ID:]
 [last_CSeq:]
-Contact: <sip:sipp@[local_ip]:[local_port];transport=[transport]>
+Contact: <sip:[$local_addr];transport=[transport]>
 Max-Forwards: 100
 User-Agent: SIPp/sippy_cup
 Content-Length: 0
 [routes]
 ]]>
 </send>
+  <Reference variables="remote_addr,local_addr,call_addr,dummy,remote_tag"/>
 </scenario>
     END
     end
 
     context "with a valid steps definition" do
-      let(:steps) { ['invite', 'wait_for_answer', 'ack_answer', 'wait_for_hangup'] }
+      let(:steps) { ['invite', 'wait_for_answer', 'wait_for_hangup'] }
 
       it "runs each step" do
         subject.build(steps)
         subject.to_xml(:pcap_path => "/dev/null").should == scenario_xml
+      end
+    end
+
+    context "having steps with arguments" do
+      let(:steps) do
+        [
+          %q(register 'user@domain.com' "my password has spaces"),
+          %q(sleep 3),
+          %q(send_digits 12345)
+        ]
+      end
+
+      it "each method should receive the correct arguments" do
+        subject.should_receive(:register).once.ordered.with('user@domain.com', 'my password has spaces')
+        subject.should_receive(:sleep).once.ordered.with('3')
+        subject.should_receive(:send_digits).once.ordered.with('12345')
+        subject.build steps
       end
     end
 
@@ -696,7 +690,6 @@ from_user: #{specs_from}
 steps:
   - invite
   - wait_for_answer
-  - ack_answer
   - sleep 3
   - send_digits '3125551234'
   - sleep 5
@@ -712,7 +705,7 @@ steps:
 <![CDATA[
 INVITE sip:[service]@[remote_ip]:[remote_port] SIP/2.0
 Via: SIP/2.0/[transport] [local_ip]:[local_port];branch=[branch]
-From: "#{specs_from}" <sip:#{specs_from}@[local_ip]>;tag=[call_number]
+From: "#{specs_from}" <sip:#{specs_from}@[local_ip]:[local_port]>;tag=[call_number]
 To: <sip:[service]@[remote_ip]:[remote_port]>
 Call-ID: [call_id]
 CSeq: [cseq] INVITE
@@ -732,20 +725,24 @@ a=rtpmap:0 PCMU/8000
 a=rtpmap:101 telephone-event/8000
 a=fmtp:101 0-15
 ]]>
-</send>
+<action><assignstr assign_to="remote_addr" value="[service]@[remote_ip]:[remote_port]"/><assignstr assign_to="local_addr" value="#{specs_from}@[local_ip]:[local_port]"/><assignstr assign_to="call_addr" value="[service]@[remote_ip]:[remote_port]"/></action></send>
   <recv response="100" optional="true"/>
   <recv response="180" optional="true"/>
   <recv response="183" optional="true"/>
-  <recv response="200" rrs="true" rtd="true"/>
+  <recv response="200" rrs="true" rtd="true">
+    <action>
+      <ereg regexp="&lt;sip:(.*)&gt;.*;tag=([^;]*)" search_in="hdr" header="To:" assign_to="dummy,remote_addr,remote_tag"/>
+    </action>
+  </recv>
   <send>
 <![CDATA[
 ACK [next_url] SIP/2.0
 Via: SIP/2.0/[transport] [local_ip]:[local_port];branch=[branch]
-From: "#{specs_from}" <sip:#{specs_from}@[local_ip]>;tag=[call_number]
+From: "#{specs_from}" <sip:#{specs_from}@[local_ip]:[local_port]>;tag=[call_number]
 To: <sip:[service]@[remote_ip]:[remote_port]>[peer_tag_param]
 Call-ID: [call_id]
 CSeq: [cseq] ACK
-Contact: <sip:#{specs_from}@[local_ip]:[local_port];transport=[transport]>
+Contact: <sip:[$local_addr];transport=[transport]>
 Max-Forwards: 100
 User-Agent: SIPp/sippy_cup
 Content-Length: 0
@@ -770,13 +767,14 @@ SIP/2.0 200 OK
 [last_To:]
 [last_Call-ID:]
 [last_CSeq:]
-Contact: <sip:#{specs_from}@[local_ip]:[local_port];transport=[transport]>
+Contact: <sip:[$local_addr];transport=[transport]>
 Max-Forwards: 100
 User-Agent: SIPp/sippy_cup
 Content-Length: 0
 [routes]
 ]]>
 </send>
+  <Reference variables="remote_addr,local_addr,call_addr,dummy,remote_tag"/>
 </scenario>
       END
     end
@@ -860,7 +858,6 @@ from_user: #{specs_from}
 steps:
 - invite
 - wait_for_answer
-- ack_answer
 - sleep 3
 - send_digits '3125551234'
 - sleep 5
@@ -894,7 +891,6 @@ from_user: #{specs_from}
 steps:
   - invite
   - wait_for_answer
-  - ack_answer
   - sleep 3
   - send_digits '3125551234'
   - sleep 5
@@ -944,9 +940,8 @@ from_user: #{specs_from}
 steps:
   - invite
   - wait_for_answer
-  - ack_answer
   - sleep 3
-  - send_digits 'abc'
+  - send_digits 'xyz'
   - sleep 5
   - send_digits '#'
   - wait_for_hangup
@@ -964,7 +959,7 @@ steps:
 
       it "sets the error messages for the scenario" do
         scenario = SippyCup::Scenario.from_manifest(scenario_yaml)
-        scenario.errors.should == [{step: 5, message: "send_digits 'abc': Invalid DTMF digit requested: a"}]
+        scenario.errors.should == [{step: 4, message: "send_digits 'xyz': Invalid DTMF digit requested: x"}]
       end
     end
   end